+++ edaflow/__init__.py
@@ -2,7 +2,19 @@
 edaflow - A Python package for exploratory data analysis workflows
 """
 
-from .analysis import check_null_columns, analyze_categorical_columns, convert_to_numeric, visualize_categorical_values, display_column_types, impute_numerical_median, impute_categorical_mode, visualize_numerical_boxplots, handle_outliers_median, visualize_interactive_boxplots
+from .analysis import (
+    check_null_columns,
+    analyze_categorical_columns,
+    convert_to_numeric,
+    visualize_categorical_values,
+    display_column_types,
+    impute_numerical_median,
+    impute_categorical_mode,
+    visualize_numerical_boxplots,
+    handle_outliers_median,
+    visualize_interactive_boxplots,
+    visualize_heatmap
+)
 
 __version__ = "0.5.1"
 __author__ = "Evan Low"